seperate network peer code to two files
